22.12.16 구글 연동 테스트 전 버전 병합

+++ Prototype/Assets/02_Script/Play/Unit/Unit.cs
@@ -52,7 +52,7 @@ public class Unit : MonoBehaviour
     [SerializeField] private ParticleSystem shieldEffect;
     [SerializeField] private ParticleSystem rageEffect;
 
-    public float skillCondition;
+    public int skillCondition;
 
     private SkeletonAnimation skeletonAnimation;
     public EUnitType UnitType { get => eUnitType; }
